Add in required files for Windows compilation/execution

+++ src/new_client/TickBuffer.cpp
@@ -419,25 +419,23 @@ void TickBuffer::sv_setchar_skill( const unsigned char* buf )
   playerData_.getClientSidePlayerInfo().skill[ n ][ 4 ] = buf[ 6 ];
   playerData_.getClientSidePlayerInfo().skill[ n ][ 5 ] = buf[ 7 ];
 
-  std::sort( playerData_.getSkillList(), playerData_.getSkillList() + MAX_SKILLS,
-             []( const skilltab& lhs, const skilltab& rhs )
-             {
-               // Begin server message processing implementation
-               int m1 = lhs.nr;
-               int m2 = rhs.nr;
+  std::sort( playerData_.getSkillList(), playerData_.getSkillList() + MAX_SKILLS, []( const skilltab& lhs, const skilltab& rhs ) {
+    // Begin server message processing implementation
+    int m1 = lhs.nr;
+    int m2 = rhs.nr;
 
-               if ( m1 == 99 && m2 != 99 )
-                 return true;
-               if ( m2 == 99 && m1 != 99 )
-                 return false;
+    if ( m1 == 99 && m2 != 99 )
+      return true;
+    if ( m2 == 99 && m1 != 99 )
+      return false;
 
-               if ( lhs.sortkey > rhs.sortkey )
-                 return true;
-               if ( lhs.sortkey < rhs.sortkey )
-                 return false;
+    if ( lhs.sortkey > rhs.sortkey )
+      return true;
+    if ( lhs.sortkey < rhs.sortkey )
+      return false;
 
-               return strcmp( lhs.name, rhs.name ) > 0;
-             } );
+    return strcmp( lhs.name, rhs.name ) > 0;
+  } );
 }
 
 void TickBuffer::sv_setchar_ahp( const unsigned char* buf )
@@ -913,13 +911,13 @@ int TickBuffer::sv_ignore( const unsigned char* buf )
 
   if ( ! start )
   {
-    start = time( NULL );
+    start = static_cast< int >( time( NULL ) );
   }
 
   if ( cnt++ > 16 )
   {
     cnt = 0;
-    d   = time( NULL ) - start;
+    d   = static_cast< int >( time( NULL ) - start );
     if ( d == 0 )
     {
       d = 1;
